+++ lib/base_lib_pub.dart
@@ -1,6 +1,8 @@
 library base_lib_pub;
 
 export 'src/base_main.dart';
+export 'src/component/image/my_extended_image_provider_with_custom_http_client.dart';
+export 'src/component/image/my_extended_image_with_http_client.dart';
 export 'src/const/global_const.dart';
 export 'src/extension/common_ext.dart';
 export 'src/extension/datetime_ext.dart';